+++ data/duke.txt
@@ -1,1 +1,1 @@
-EVENT | 0 | class | 1pm | 4pm
+TODO | 0 | hihi
