+++ text-ui-test/input.txt
@@ -1,15 +1,1 @@
-hi
-todo write essays
-event interviews /from today 2pm /to 3pm
-deadline project /by 2020-02-10 morning
-check list
-mark 3
-mark 2
-unmark 3
-mark 8
-todo
-event test /from test to
-deadline test /by
-delete 1
-list
 88
